+++ src/lib/AsyncTypeaheadInput.tsx
@@ -195,7 +195,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadInputPr
           const values = convertAutoCompleteOptionsToStringArray(optionsArray);
           const finalValue = multiple ? values : values[0];
           if (onChange) {
-            onChange(finalValue);
+            void (async () => await onChange(finalValue))();
           }
           field.onChange(finalValue);
         }}
@@ -207,7 +207,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadInputPr
             setDebounceSearch({ delay, query });
           }
           if (onInputChange) {
-            onInputChange(query, reason);
+            void (async () => await onInputChange(query, reason))();
           }
         }}
         renderOption={highlightOptions ? renderHighlightedOptionFunction : undefined}