+++ src/lib/types/Typeahead.ts
@@ -29,8 +29,8 @@ interface CommonTypeaheadProps<T extends FieldValues> extends Omit<
   innerRef?: RefObject<HTMLInputElement | null>;
   fitMenuContent?: boolean;
   getOptionDisabled?: (option: TypeaheadOption) => boolean;
-  onChange?: (selected: string | string[]) => void;
-  onInputChange?: (text: string, reason: AutocompleteInputChangeReason) => void;
+  onChange?: (selected: string | string[]) => void | Promise<void>;
+  onInputChange?: (text: string, reason: AutocompleteInputChangeReason) => void | Promise<void>;
   onClose?: (event: SyntheticEvent, reason: AutocompleteCloseReason) => void;
   onOpen?: (event: SyntheticEvent) => void;
   onBlur?: () => void;